requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
face_recognition>=1.3.0
numpy>=1.24
Pillow>=9.0
werkzeug>=2.0
python-dotenv>=1.0
